full credit assignment about 7 times faster
To simplify management of merges and new heuristics the application
just wipes the join model, computes the new names, and reassings
commits to contributors according to the new ones or rules.

This is fast enough and much easy to understand to trying to be
too clever. The price is filling the join table, which as of this
writing has 62K rows.

Doing a bulk import with COPY FROM pays off here, we go from about
3 minutes, to about 25 seconds.