Count contribution toward committers too.
If a commit was written by an author but committed by someone else,
the contributors app currently generates no contribution for the committer.

Which is not what we want because the committer is one from the
Rails team who won't have a due contribution counted their way.

These cases happen if the committer manually squashes a PR, for instance,
through GitHubs squash commits merge button.

Fix by appending the committer name in those cases.