couple of stylistic details
The current style of the project is single quotes unless double
are needed. This is no longer the Rails style, but if we change
it will be done for the entire project.

On the other hand, the test suite of canonical names puts the
name to be mapped in the test description, and the canonical
name in the assertion.