Make website responsive
This uses CSS and a bit of JS to make the website responsive. Some CSS was
updated to more modern version. For instance, images were used to add a shadow
on the main content. That was replaced by a box-shadow. The navigation element
is inside a details/summary element, so it can be toggled on mobile devices to
save space. It is closed by default on mobile devices, but open on desktops.